refactor(dropper)!: fix re-exports, centralize types
`async_dropper` has been re-exporting the derive macro somewhat
unoptimally, and as a result the types that need to be shared across
both implementations were not properly shared.

In addition, Traits being defined in the derive macro *should not*
have been, as they were causing issues when trying to use multiple
derive macros in the same codebase.

This commit re-structures the async_dropper crate by moving the types
expected and used by derive and simple into `async_dropper` itself, and
creating features for derive and simple to be used.

3.x versions of `async-dropper-derive` (which is re-exported through
this crate) *require* this change, which should be present in version
3.x and higher of `async-dropper`.

Signed-off-by: vados <vados@vadosware.io>