fix(simple): timeout usage
The previous implementation of timeouts for dropping had a bug -- the
timeout value was saved *after* calling `std::mem::take`, and was
referring to the new `T::default()` instance's value of `timeout`,
which is always `None`.

This commit fixes the timeout usage by taking the timeout value before
`std::mem::take`.

Co-authored-by: Linken Quy Dinh <linken.dinh@gmail.com>